+++ lib/modbot/reddit_wrap.rb
@@ -108,6 +108,7 @@ module RedditWrap
           h.id = yy['data']['id']
           h.fullid = yy['data']['name']
           h.permalink = reddit_route(yy['data']['permalink'])
+          #@minimal_author  ?  h.author = yy['data']['author'] : reddit_user(yy['data']['author'])
           h.author = reddit_user(yy['data']['author'])
           if yy['kind'] == "t1"
             h.kind = "comment"